modified the create exercise alert dialog to accept decimal numbers for weight

+++ Application_codebase/app/src/main/java/com/example/workouttrackerapplication/ui/create_workout/CreateExerciseDialogFragment.java
@@ -37,7 +37,7 @@ public class CreateExerciseDialogFragment extends AppCompatDialogFragment {
                                 exName.getText().toString(),
                                 Integer.parseInt(sets.getText().toString()),
                                 Integer.parseInt(reps.getText().toString()),
-                                Integer.parseInt(weight.getText().toString()));
+                                Double.parseDouble(weight.getText().toString()));
                         CreateWorkoutFragment.displayList.add(exerciseModel);
                     }
                     catch (Exception e){
